+++ .gitignore
@@ -8,3 +8,4 @@ wandb/*
 models/*
 documentation/files
 midi_files/*
+upload_models.py
